Fix http already bind address problem in tests
In the one hand i was registering two times the before/after scenario
callbacks. In the other hand, rununtil.Killed() seems to be not a
properly solution for bringing up and down the test http server, since
CancelFunc of built in context.go will not wait for http server
termination. This is a problem if you are doing multiple scenario testing.
Now it will wait for http shutdown after each scenario.